Move the execute method of the bootloader command
Move the code to pyanaconda.bootloader.execution.
